refactor: replace respond.js with reply.js for message handling
